Prioritize stats relating to the router.
Now that there's a single deque per destination, we can put router
stats into the deque on the left.  To do that, when a router is
initialized the metricReceived and the metricGenerated callbacks are
different - they're now pointed at the
client_manager.sendHighPriorityDatapoint method, which will push onto
the left of the deque.

The goal here is that when there's a failure due to the router being
overwhelmed, the routers own stats tell that story.  But currently
they're not likely to be among the first stats pushed out, and so
they'll be lost.  The new sendHighPriorityDatapoint method will ignore
the maximum queue size, and it'll cause these stats to go out first in
any batch.