When delivering metrics to a destination a long way away, the relay can run into poor internet weather after a time.
This patch checks for the quality of the connection in a given
CARBON_METRIC_INTERVAL.  If the number of metrics received in the
interval is greater than MIN_RESET_STAT_FLOW, and the ratio of metrics
received to metrics sent falls below MIN_RESET_RATIO, then the
connection will be re-set.  This behavior can be enabled by setting
USE_RATIO_RESET=True in the config for the relay.

This will only work for a setup that expects all inbound stats to be
sent outbound as well.  A configuration where stats are filtered would
require that an extra relay is set up on the same host or on a very
nearby host that will act as the quality monitor relay.  The
first-layer relay can filter the stats, the second-layer relay can
then forward all stats it receives and monitor the connection.

Bugs fixed:

- In some places in client.py I had been referring to "settings" as
  "config" because of a typo.

- The specialMetricGenerated Event wasn't sending events.  I've
  reverted to using the regular metricGenerated Event until I can
  figure out what I'm missing.

Conflicts:
	lib/carbon/conf.py